# improve CLI import performance