--- a/src/prefect/_internal/compatibility/deprecated.py
+++ b/src/prefect/_internal/compatibility/deprecated.py
@@ -18,7 +18,6 @@ import sys
 import warnings
 from typing import TYPE_CHECKING, Any, Callable, Optional, Union
 
-import dateparser
 from pydantic import BaseModel
 from typing_extensions import ParamSpec, TypeAlias, TypeVar
 
@@ -64,6 +63,8 @@ def _coerce_datetime(
         return dt
     with warnings.catch_warnings():
         warnings.filterwarnings("ignore", category=DeprecationWarning)
+        import dateparser
+
         return dateparser.parse(dt)
 
 
@@ -106,15 +107,14 @@ def deprecated_callable(
     help: str = "",
 ) -> Callable[[Callable[P, R]], Callable[P, R]]:
     def decorator(fn: Callable[P, R]) -> Callable[P, R]:
-        message = generate_deprecation_message(
-            name=to_qualified_name(fn),
-            start_date=start_date,
-            end_date=end_date,
-            help=help,
-        )
-
         @functools.wraps(fn)
         def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
+            message = generate_deprecation_message(
+                name=to_qualified_name(fn),
+                start_date=start_date,
+                end_date=end_date,
+                help=help,
+            )
             warnings.warn(message, PrefectDeprecationWarning, stacklevel=stacklevel)
             return fn(*args, **kwargs)
 
@@ -131,17 +131,16 @@ def deprecated_class(
     help: str = "",
 ) -> Callable[[type[T]], type[T]]:
     def decorator(cls: type[T]) -> type[T]:
-        message = generate_deprecation_message(
-            name=to_qualified_name(cls),
-            start_date=start_date,
-            end_date=end_date,
-            help=help,
-        )
-
         original_init = cls.__init__
 
         @functools.wraps(original_init)
         def new_init(self: T, *args: Any, **kwargs: Any) -> None:
+            message = generate_deprecation_message(
+                name=to_qualified_name(cls),
+                start_date=start_date,
+                end_date=end_date,
+                help=help,
+            )
             warnings.warn(message, PrefectDeprecationWarning, stacklevel=stacklevel)
             original_init(self, *args, **kwargs)
 
@@ -177,14 +176,6 @@ def deprecated_parameter(
     when = when or (lambda _: True)
 
     def decorator(fn: Callable[P, R]) -> Callable[P, R]:
-        message = generate_deprecation_message(
-            name=f"The parameter {name!r} for {fn.__name__!r}",
-            start_date=start_date,
-            end_date=end_date,
-            help=help,
-            when=when_message,
-        )
-
         @functools.wraps(fn)
         def wrapper(*args: P.args, **kwargs: P.kwargs) -> R:
             try:
@@ -194,6 +185,13 @@ def deprecated_parameter(
                 parameters = kwargs
 
             if name in parameters and when(parameters[name]):
+                message = generate_deprecation_message(
+                    name=f"The parameter {name!r} for {fn.__name__!r}",
+                    start_date=start_date,
+                    end_date=end_date,
+                    help=help,
+                    when=when_message,
+                )
                 warnings.warn(message, PrefectDeprecationWarning, stacklevel=stacklevel)
 
             return fn(*args, **kwargs)
@@ -238,19 +236,18 @@ def deprecated_field(
     # Replaces the model's __init__ method with one that performs an additional warning
     # check
     def decorator(model_cls: type[M]) -> type[M]:
-        message = generate_deprecation_message(
-            name=f"The field {name!r} in {model_cls.__name__!r}",
-            start_date=start_date,
-            end_date=end_date,
-            help=help,
-            when=when_message,
-        )
-
         cls_init = model_cls.__init__
 
         @functools.wraps(model_cls.__init__)
         def __init__(__pydantic_self__: M, **data: Any) -> None:
             if name in data.keys() and when(data[name]):
+                message = generate_deprecation_message(
+                    name=f"The field {name!r} in {model_cls.__name__!r}",
+                    start_date=start_date,
+                    end_date=end_date,
+                    help=help,
+                    when=when_message,
+                )
                 warnings.warn(message, PrefectDeprecationWarning, stacklevel=stacklevel)
 
             cls_init(__pydantic_self__, **data)